backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
celery[redis]==5.3.4
redis==4.6.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
pyyaml==6.0.1

# Image hashing dependencies
imagehash==4.3.1
Pillow==10.1.0
face_recognition==1.3.0
numpy==1.24.3

# Web scraping dependencies
playwright==1.40.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
